requirements.txt
rasa==3.6.20
jieba==0.42.1
transformers==4.41.1
mysql-connector-python==8.4.0
